backend/requirements.txt
dateutils==0.6.11
Django==3.1.1
django-cors==0.1
djangorestframework==3.11.1
python-dateutil~=2.8.1